+++ bin/stress-test/src/seeding/mod.rs
@@ -9,6 +9,7 @@ use miden_block_prover::LocalBlockProver;
 use miden_lib::account::auth::AuthRpoFalcon512;
 use miden_lib::account::faucets::BasicFungibleFaucet;
 use miden_lib::account::wallets::BasicWallet;
+use miden_lib::block::build_block;
 use miden_lib::note::create_p2id_note;
 use miden_lib::utils::Serializable;
 use miden_node_block_producer::store::StoreClient;
@@ -245,8 +246,12 @@ async fn apply_block(
     store_client: &StoreClient,
     metrics: &mut SeedingMetrics,
 ) -> ProvenBlock {
-    let proposed_block = ProposedBlock::new(block_inputs, batches).unwrap();
-    let proven_block = LocalBlockProver::new(0).prove_dummy(proposed_block).unwrap();
+    let proposed_block = ProposedBlock::new(block_inputs.clone(), batches).unwrap();
+    let (header, body) = build_block(proposed_block.clone()).unwrap();
+    let block_proof = LocalBlockProver::new(0)
+        .prove_dummy(proposed_block.batches().clone(), header.clone(), block_inputs)
+        .unwrap();
+    let proven_block = ProvenBlock::new_unchecked(header, body, block_proof);
     let block_size: usize = proven_block.to_bytes().len();
 
     let start = Instant::now();