+++ crates/block-producer/src/block_builder/mod.rs
@@ -6,10 +6,18 @@ use futures::never::Never;
 use miden_block_prover::LocalBlockProver;
 use miden_node_utils::tracing::OpenTelemetrySpanExt;
 use miden_objects::MIN_PROOF_SECURITY_LEVEL;
-use miden_objects::batch::ProvenBatch;
-use miden_objects::block::{BlockInputs, BlockNumber, ProposedBlock, ProvenBlock};
+use miden_objects::batch::{OrderedBatches, ProvenBatch};
+use miden_objects::block::{
+    BlockBody,
+    BlockHeader,
+    BlockInputs,
+    BlockNumber,
+    BlockProof,
+    ProposedBlock,
+    ProvenBlock,
+};
 use miden_objects::note::NoteHeader;
-use miden_objects::transaction::TransactionHeader;
+use miden_objects::transaction::{OrderedTransactionHeaders, TransactionHeader};
 use miden_remote_prover_client::remote_prover::block_prover::RemoteBlockProver;
 use rand::Rng;
 use tokio::time::Duration;
@@ -19,6 +27,7 @@ use url::Url;
 use crate::errors::BuildBlockError;
 use crate::mempool::SharedMempool;
 use crate::store::StoreClient;
+use crate::validator::BlockProducerValidatorClient;
 use crate::{COMPONENT, TelemetryInjectorExt};
 
 // BLOCK BUILDER
@@ -36,6 +45,8 @@ pub struct BlockBuilder {
 
     pub store: StoreClient,
 
+    pub validator: BlockProducerValidatorClient,
+
     /// The prover used to prove a proposed block into a proven block.
     pub block_prover: BlockProver,
 }
@@ -46,6 +57,7 @@ impl BlockBuilder {
     /// If the block prover URL is not set, the block builder will use the local block prover.
     pub fn new(
         store: StoreClient,
+        validator: BlockProducerValidatorClient,
         block_prover_url: Option<Url>,
         block_interval: Duration,
     ) -> Self {
@@ -61,6 +73,7 @@ impl BlockBuilder {
             failure_rate: 0.0,
             block_prover,
             store,
+            validator,
         }
     }
     /// Starts the [`BlockBuilder`], infinitely producing blocks at the configured interval.
@@ -112,8 +125,12 @@ impl BlockBuilder {
         self.get_block_inputs(selected)
             .inspect_ok(BlockBatchesAndInputs::inject_telemetry)
             .and_then(|inputs| self.propose_block(inputs))
-            .inspect_ok(ProposedBlock::inject_telemetry)
-            .and_then(|inputs| self.prove_block(inputs))
+            .inspect_ok(|(proposed_block, _)| {
+                ProposedBlock::inject_telemetry(proposed_block);
+            })
+            .and_then(|(proposed_block, inputs)| self.validate_block(proposed_block, inputs))
+            .and_then(|(proposed_block, inputs, header, body)| self.prove_block(proposed_block, inputs, header, body))
+            .and_then(|(proposed_block, header, body, block_proof)| self.construct_proven_block(proposed_block, header, body, block_proof))
             .inspect_ok(ProvenBlock::inject_telemetry)
             // Failure must be injected before the final pipeline stage i.e. before commit is called. The system cannot
             // handle errors after it considers the process complete (which makes sense).
@@ -197,31 +214,68 @@ impl BlockBuilder {
     async fn propose_block(
         &self,
         batches_inputs: BlockBatchesAndInputs,
-    ) -> Result<ProposedBlock, BuildBlockError> {
+    ) -> Result<(ProposedBlock, BlockInputs), BuildBlockError> {
         let BlockBatchesAndInputs { batches, inputs } = batches_inputs;
         let batches = batches.into_iter().map(Arc::unwrap_or_clone).collect();
 
-        let proposed_block =
-            ProposedBlock::new(inputs, batches).map_err(BuildBlockError::ProposeBlockFailed)?;
+        let proposed_block = ProposedBlock::new(inputs.clone(), batches)
+            .map_err(BuildBlockError::ProposeBlockFailed)?;
 
-        Ok(proposed_block)
+        Ok((proposed_block, inputs))
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.validate_block", skip_all, err)]
+    async fn validate_block(
+        &self,
+        proposed_block: ProposedBlock,
+        block_inputs: BlockInputs,
+    ) -> Result<(OrderedBatches, BlockInputs, BlockHeader, BlockBody), BuildBlockError> {
+        let response = self
+            .validator
+            .validate_block(proposed_block.clone())
+            .await
+            .map_err(BuildBlockError::ValidateBlockFailed)?;
+
+        // TODO: Check that the returned header and body match the proposed block.
+
+        let (ordered_batches, ..) = proposed_block.into_parts();
+        Ok((ordered_batches, block_inputs, response.header, response.body))
     }
 
     #[instrument(target = COMPONENT, name = "block_builder.prove_block", skip_all, err)]
     async fn prove_block(
         &self,
-        proposed_block: ProposedBlock,
-    ) -> Result<ProvenBlock, BuildBlockError> {
-        let proven_block = self.block_prover.prove(proposed_block).await?;
+        ordered_batches: OrderedBatches,
+        block_inputs: BlockInputs,
+        header: BlockHeader,
+        body: BlockBody,
+    ) -> Result<(OrderedBatches, BlockHeader, BlockBody, BlockProof), BuildBlockError> {
+        // Prove block using header and body from validator.
+        let block_proof = self
+            .block_prover
+            .prove(ordered_batches.clone(), header.clone(), block_inputs)
+            .await?;
+        self.simulate_proving().await;
 
+        Ok((ordered_batches, header, body, block_proof))
+    }
+
+    #[instrument(target = COMPONENT, name = "block_builder.construct_proven_block", skip_all, err)]
+    async fn construct_proven_block(
+        &self,
+        ordered_batches: OrderedBatches,
+        header: BlockHeader,
+        body: BlockBody,
+        block_proof: BlockProof,
+    ) -> Result<ProvenBlock, BuildBlockError> {
+        let proven_block = ProvenBlock::new_unchecked(header, body, block_proof);
         if proven_block.proof_security_level() < MIN_PROOF_SECURITY_LEVEL {
             return Err(BuildBlockError::SecurityLevelTooLow(
                 proven_block.proof_security_level(),
                 MIN_PROOF_SECURITY_LEVEL,
             ));
         }
-
-        self.simulate_proving().await;
+        validate_tx_headers(&proven_block, &ordered_batches.to_transactions())?;
 
         Ok(proven_block)
     }
@@ -388,15 +442,55 @@ impl BlockProver {
     }
 
     #[instrument(target = COMPONENT, skip_all, err)]
-    async fn prove(&self, proposed_block: ProposedBlock) -> Result<ProvenBlock, BuildBlockError> {
+    async fn prove(
+        &self,
+        tx_batches: OrderedBatches,
+        block_header: BlockHeader,
+        block_inputs: BlockInputs,
+    ) -> Result<BlockProof, BuildBlockError> {
         match self {
-            Self::Local(prover) => {
-                prover.prove(proposed_block).map_err(BuildBlockError::ProveBlockFailed)
-            },
+            Self::Local(prover) => prover
+                .prove(tx_batches, block_header, block_inputs)
+                .map_err(BuildBlockError::ProveBlockFailed),
             Self::Remote(prover) => prover
-                .prove(proposed_block)
+                .prove(tx_batches, block_header, block_inputs)
                 .await
                 .map_err(BuildBlockError::RemoteProverClientError),
         }
     }
+}
+
+/// Validates that the proven block's transaction headers are consistent with the transactions
+/// passed in the proposed block.
+///
+/// This expects that transactions from the proposed block and proven block are in the same
+/// order, as define by [`OrderedTransactionHeaders`].
+fn validate_tx_headers(
+    proven_block: &ProvenBlock,
+    proposed_txs: &OrderedTransactionHeaders,
+) -> Result<(), BuildBlockError> {
+    if proposed_txs.as_slice().len() != proven_block.body().transactions().as_slice().len() {
+        return Err(BuildBlockError::other(format!(
+            "remote prover returned {} transaction headers but {} transactions were passed as part of the proposed block",
+            proven_block.body().transactions().as_slice().len(),
+            proposed_txs.as_slice().len()
+        )));
+    }
+
+    // Because we checked the length matches we can zip the iterators up.
+    // We expect the transaction headers to be in the same order.
+    for (proposed_header, proven_header) in proposed_txs
+        .as_slice()
+        .iter()
+        .zip(proven_block.body().transactions().as_slice())
+    {
+        if proposed_header != proven_header {
+            return Err(BuildBlockError::other(format!(
+                "transaction header with id {} does not match header of the transaction in the proposed block",
+                proposed_header.id()
+            )));
+        }
+    }
+
+    Ok(())
 }