+++ crates/block-producer/src/errors.rs
@@ -1,4 +1,6 @@
-use miden_block_prover::ProvenBlockError;
+use core::error::Error as CoreError;
+
+use miden_block_prover::BlockProverError;
 use miden_node_proto::errors::{ConversionError, GrpcError};
 use miden_objects::account::AccountId;
 use miden_objects::block::BlockNumber;
@@ -8,6 +10,8 @@ use miden_objects::{ProposedBatchError, ProposedBlockError, ProvenBatchError, Wo
 use miden_remote_prover_client::RemoteProverClientError;
 use thiserror::Error;
 use tokio::task::JoinError;
+
+use crate::validator::ValidatorError;
 
 // Block-producer errors
 // =================================================================================================
@@ -204,8 +208,10 @@ pub enum BuildBlockError {
     GetBlockInputsFailed(#[source] StoreError),
     #[error("failed to propose block")]
     ProposeBlockFailed(#[source] ProposedBlockError),
+    #[error("failed to validate block")]
+    ValidateBlockFailed(#[source] ValidatorError),
     #[error("failed to prove block")]
-    ProveBlockFailed(#[source] ProvenBlockError),
+    ProveBlockFailed(#[source] BlockProverError),
     /// We sometimes randomly inject errors into the batch building process to test our failure
     /// responses.
     #[error("nothing actually went wrong, failure was injected on purpose")]
@@ -214,6 +220,21 @@ pub enum BuildBlockError {
     RemoteProverClientError(#[source] RemoteProverClientError),
     #[error("block proof security level is too low: {0} < {1}")]
     SecurityLevelTooLow(u32, u32),
+    /// Custom error variant for errors not covered by the other variants.
+    #[error("{error_msg}")]
+    Other {
+        error_msg: Box<str>,
+        source: Option<Box<dyn CoreError + Send + Sync + 'static>>,
+    },
+}
+
+impl BuildBlockError {
+    /// Creates a custom error using the [`BuildBlockError::Other`] variant from an
+    /// error message.
+    pub fn other(message: impl Into<String>) -> Self {
+        let message: String = message.into();
+        Self::Other { error_msg: message.into(), source: None }
+    }
 }
 
 // Store errors