+++ crates/block-producer/src/lib.rs
@@ -9,6 +9,7 @@ mod block_builder;
 mod domain;
 mod mempool;
 pub mod store;
+pub mod validator;
 
 #[cfg(feature = "testing")]
 pub mod errors;