+++ crates/remote-prover-client/src/remote_prover/block_prover.rs
@@ -3,10 +3,12 @@ use alloc::sync::Arc;
 use alloc::vec::Vec;
 use core::time::Duration;
 
-use miden_objects::batch::ProvenBatch;
-use miden_objects::block::{ProposedBlock, ProvenBlock};
+use miden_node_proto::BlockProofRequest;
+use miden_objects::batch::{OrderedBatches, ProvenBatch};
+use miden_objects::block::{BlockHeader, BlockInputs, BlockProof, ProposedBlock, ProvenBlock};
 use miden_objects::transaction::{OrderedTransactionHeaders, TransactionHeader};
 use miden_objects::utils::{Deserializable, DeserializationError, Serializable};
+use miden_tx::utils::{ByteReader, ByteWriter};
 use tokio::sync::Mutex;
 
 use super::generated::api_client::ApiClient;
@@ -104,8 +106,10 @@ impl RemoteBlockProver {
 impl RemoteBlockProver {
     pub async fn prove(
         &self,
-        proposed_block: ProposedBlock,
-    ) -> Result<ProvenBlock, RemoteProverClientError> {
+        tx_batches: OrderedBatches,
+        block_header: BlockHeader,
+        block_inputs: BlockInputs,
+    ) -> Result<BlockProof, RemoteProverClientError> {
         use miden_objects::utils::Serializable;
         self.connect().await?;
 
@@ -119,75 +123,38 @@ impl RemoteBlockProver {
             })?
             .clone();
 
-        // Get the set of expected transaction headers.
-        let proposed_txs = proposed_block.batches().to_transactions();
-
-        let request = tonic::Request::new(proposed_block.into());
+        let request = BlockProofRequest { tx_batches, block_header, block_inputs };
+        let request = tonic::Request::new(request.into());
 
         let response = client.prove(request).await.map_err(|err| {
             RemoteProverClientError::other_with_source("failed to prove block", err)
         })?;
 
-        // Deserialize the response bytes back into a ProvenBlock.
-        let proven_block = ProvenBlock::try_from(response.into_inner()).map_err(|err| {
+        // Deserialize the response bytes back into a BlockProof.
+        let block_proof = BlockProof::try_from(response.into_inner()).map_err(|err| {
             RemoteProverClientError::other_with_source(
                 "failed to deserialize received response from remote block prover",
                 err,
             )
         })?;
 
-        Self::validate_tx_headers(&proven_block, &proposed_txs)?;
-
-        Ok(proven_block)
-    }
-
-    /// Validates that the proven block's transaction headers are consistent with the transactions
-    /// passed in the proposed block.
-    ///
-    /// This expects that transactions from the proposed block and proven block are in the same
-    /// order, as define by [`OrderedTransactionHeaders`].
-    fn validate_tx_headers(
-        proven_block: &ProvenBlock,
-        proposed_txs: &OrderedTransactionHeaders,
-    ) -> Result<(), RemoteProverClientError> {
-        if proposed_txs.as_slice().len() != proven_block.transactions().as_slice().len() {
-            return Err(RemoteProverClientError::other(format!(
-                "remote prover returned {} transaction headers but {} transactions were passed as part of the proposed block",
-                proven_block.transactions().as_slice().len(),
-                proposed_txs.as_slice().len()
-            )));
-        }
-
-        // Because we checked the length matches we can zip the iterators up.
-        // We expect the transaction headers to be in the same order.
-        for (proposed_header, proven_header) in
-            proposed_txs.as_slice().iter().zip(proven_block.transactions().as_slice())
-        {
-            if proposed_header != proven_header {
-                return Err(RemoteProverClientError::other(format!(
-                    "transaction header with id {} does not match header of the transaction in the proposed block",
-                    proposed_header.id()
-                )));
-            }
-        }
-
-        Ok(())
+        Ok(block_proof)
     }
 }
 
 // CONVERSION
 // ================================================================================================
 
-impl TryFrom<proto::Proof> for ProvenBlock {
+impl TryFrom<proto::Proof> for BlockProof {
     type Error = DeserializationError;
 
     fn try_from(value: proto::Proof) -> Result<Self, Self::Error> {
-        ProvenBlock::read_from_bytes(&value.payload)
+        BlockProof::read_from_bytes(&value.payload)
     }
 }
 
-impl From<ProposedBlock> for proto::ProofRequest {
-    fn from(proposed_block: ProposedBlock) -> Self {
+impl From<BlockProofRequest> for proto::ProofRequest {
+    fn from(proposed_block: BlockProofRequest) -> Self {
         proto::ProofRequest {
             proof_type: proto::ProofType::Block.into(),
             payload: proposed_block.to_bytes(),